+++ src/util.ts
@@ -266,21 +266,22 @@ export interface RopmPackageJson {
     files?: string[];
     keywords?: string[];
     version: string;
-    ropm?: {
-        /**
-         * The path to the rootDir of the project where `ropm` should install all ropm modules
-         */
-        rootDir?: string;
-        /**
-         * The path to the rootDir of the a ropm module's package files. Use this if your module stores files in a subdirectory.
-         * NOTE: This should only be used by ropm package AUTHORS
-         */
-        packageRootDir?: string;
-        /**
-         * An array of module aliases that should not be prefixed when installed into `rootDir`. Use this with caution.
-         */
-        noprefix?: string[];
-    };
+    ropm?: RopmOptions;
+}
+export interface RopmOptions {
+    /**
+     * The path to the rootDir of the project where `ropm` should install all ropm modules
+     */
+    rootDir?: string;
+    /**
+     * The path to the rootDir of the a ropm module's package files. Use this if your module stores files in a subdirectory.
+     * NOTE: This should only be used by ropm package AUTHORS
+     */
+    packageRootDir?: string;
+    /**
+     * An array of module aliases that should not be prefixed when installed into `rootDir`. Use this with caution.
+     */
+    noprefix?: string[];
 }
 
 export interface ModuleDependency {