+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+## [0.6.0]
+### Added
+ - Find and prefix all function identifiers across the entire package.
+
+
+
 ## [0.5.0]
 ### Added
  - `ropm copy` command that speeds up local development by skipping the registry check (i.e. internal `npm install` call)
@@ -88,3 +94,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.4.0]:   https://github.com/rokucommunity/ropm/compare/v0.3.0...v0.4.0
 [0.4.1]:   https://github.com/rokucommunity/ropm/compare/v0.4.0...v0.4.1
 [0.5.0]:   https://github.com/rokucommunity/ropm/compare/v0.4.1...v0.5.0
+[0.6.0]:   https://github.com/rokucommunity/ropm/compare/v0.5.0...v0.6.0